% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopr.R
\name{chopr}
\alias{chopr}
\title{Helper for truncating probabilities}
\usage{
chopr(x, probs = c(0.025, 0.975))
}
\arguments{
\item{x}{numeric vector}

\item{probs}{vector of probabilities}
}
\value{
numeric vector
}
\description{
Helper for truncating probabilities
}
