% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{:=}
\alias{:=}
\title{Export data.table infix operator for use within package}
\usage{
`:=`(...)
}
\arguments{
\item{...}{arguments defined for use in j of DT[i, j, k]}
}
\description{
Export data.table infix operator for use within package
}
