% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_analysis.R
\name{create_formula}
\alias{create_formula}
\title{Helper: create formula for analysis}
\usage{
create_formula(outcome, exposure, covariates = NULL)
}
\arguments{
\item{outcome}{name of outcome variable}

\item{exposure}{name of exposure variable}

\item{covariates}{vector of covariate names}
}
\value{
formula for analysis
}
\description{
Helper: create formula for analysis
}
