% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{createStateArray}
\alias{createStateArray}
\title{Create a State Array from a Formula and Covariate Data}
\usage{
createStateArray(formula, covars)
}
\arguments{
\item{formula}{A formula object representing the covariates to be used in the state array.}

\item{covars}{A data frame containing the covariates. The column names of the data frame
should match the terms in the formula.}
}
\value{
An array where the dimensions correspond to the unique values of the covariates
specified in the formula. If there is one covariate term, the function returns a 2D array (NxN),
and if there are multiple covariates, it returns a multi-dimensional array.
}
\description{
This function takes a formula and a data frame of covariates to generate
an array of possible states for a hidden Markov model. The dimensions of
the array correspond to the unique values of the covariates in the formula.
}
\examples{
# Example data
state_vars <- data.frame(LH = rep(c("Sub", "Yr", "Dead"),2),
                     loc = rep(c("A", "B", "C"),2))

state_vars$LH <- factor(state_vars$LH, levels = c("Sub", "Yr", "Dead"))
state_vars$loc <- factor(state_vars$loc, levels = c("C", "A", "B"))

# Example formula
frm <- formula('~ LH + loc')

# Create the state array
state_array <- createStateArray(frm, state_vars)

}
