% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfit_lasso_weights.R
\name{crossfit_lasso_weights}
\alias{crossfit_lasso_weights}
\title{estimate crossfit lasso-based weights}
\usage{
crossfit_lasso_weights(
  internal_data,
  external_data,
  select_vars = c("id", "internal", "age_50", "female", "nhw", "hypertension",
    "diabetes", "cancer", "anxiety", "depression", "bmi_obese", "bmi_overweight",
    "bmi_underweight", "bmi_unknown"),
  weight_outcome_var = "WTFA_A",
  samp_var = "samp_WTFA_A",
  external_dataset = "NHIS",
  dataset_name = "MGI",
  id_var = "id",
  internal_var = "internal",
  cancer_factor = FALSE,
  ncores = parallelly::availableCores()/4,
  folds = 5
)
}
\arguments{
\item{internal_data}{data.table of internal data}

\item{external_data}{data.table of external data}

\item{select_vars}{vector of variables to select}

\item{weight_outcome_var}{variable name for weight outcome}

\item{samp_var}{variable name for sampling indicator}

\item{external_dataset}{name of external dataset}

\item{dataset_name}{name of internal dataset}

\item{id_var}{name of id variable}

\item{internal_var}{name of internal dataset indicator variable}

\item{cancer_factor}{whether to use a cancer factor}

\item{ncores}{number of cores to use}

\item{folds}{number of folds to use}
}
\value{
return table of results from model for the exposures
}
\description{
estimate crossfit lasso-based weights
}
