% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0.R
\name{data0}
\alias{data0}
\title{Prep data for obtaining predictions}
\usage{
data0(data, vars)
}
\arguments{
\item{data}{data.frame or data.table}

\item{vars}{character vector of variables to keep}
}
\value{
Checks to see whether packages are already installed or loaded. If not installed, will install. If already loaded, will skip loading. Otherwise, will load libraries and provide a quick summary.
}
\description{
Prep data for obtaining predictions
}
\examples{
\dontrun{
data0(mtcars, c("mpg", "cyl", "disp", "make", "model"))
}
}
