% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libri.R
\name{.check_for_deps}
\alias{.check_for_deps}
\title{Helper: Check for packages needed to use libri()}
\usage{
.check_for_deps()
}
\value{
Logical. TRUE if `cli` and `remotes` packages exist in installed packages, FALSE otherwise.
}
\description{
Helper: Check for packages needed to use libri()
}
