% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libri.R
\name{.check_library_loaded}
\alias{.check_library_loaded}
\title{Helper: Check to see if packages are loaded}
\usage{
.check_library_loaded(x)
}
\arguments{
\item{x}{Package name as string}
}
\value{
Logical. TRUE if the package exists, FALSE it not, indicating need for installation.
}
\description{
Helper: Check to see if packages are loaded
}
