% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Print data.table with nice features}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Currently unused}

\item{pkgname}{Currently unused}
}
\value{
Sets print.data.table options
}
\description{
Print data.table with nice features
}
