% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcor.R
\name{.pcor_helper}
\alias{.pcor_helper}
\title{Helper: perform rapid (parallel) partial correlation calculations on a matrix}
\usage{
.pcor_helper(x, covs, covs1 = NULL, ncores)
}
\arguments{
\item{x}{matrix of variables over which to calculate pairwise correlations}

\item{covs}{matrix of variables that serve as covariates}

\item{covs1}{alternate matrix of variables that serve as covariates when pairwise correlation is not complete}

\item{ncores}{number of cores over which to parallelize}
}
\value{
data.table with unweighted pairwise partial correlations
}
\description{
Helper: perform rapid (parallel) partial correlation calculations on a matrix
}
