% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcor.R
\name{.pwcor_helper}
\alias{.pwcor_helper}
\title{Helper: perform rapid (parallel) weighted partial correlation calculations on a matrix}
\usage{
.pwcor_helper(x, covs, w = NULL, ncores)
}
\arguments{
\item{x}{matrix of variables over which to calculate pairwise correlations}

\item{covs}{matrix of variables that serve as covariates}

\item{w}{vector of weights}

\item{ncores}{number of cores over which to parallelize}
}
\value{
data.table with weighted pairwise partial correlations
}
\description{
Helper: perform rapid (parallel) weighted partial correlation calculations on a matrix
}
