% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence_analysis.R
\name{.quick_cooccur_mod}
\alias{.quick_cooccur_mod}
\title{Helper: perform rapid unweighted and weighted cooccurrence analysis}
\usage{
.quick_cooccur_mod(
  data,
  covariates = c("age_at_threshold", "female", "length_followup"),
  outcome = "case",
  exposure = "X157",
  weight_dsn = NULL,
  evalue = TRUE
)
}
\arguments{
\item{data}{dataset}

\item{covariates}{vector of covariates}

\item{outcome}{outcome variable}

\item{exposure}{exposure variable}

\item{weight_dsn}{survey design object}

\item{evalue}{logical; calculate evalues (default = TRUE)}
}
\value{
data.table with results from cooccurrence analysis
}
\description{
Helper: perform rapid unweighted and weighted cooccurrence analysis
}
