% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_variables.R
\name{.scale_manual}
\alias{.scale_manual}
\title{Helper: mean standardize a vector}
\usage{
.scale_manual(x)
}
\arguments{
\item{x}{vector of values to standardize}
}
\value{
vector of mean standardized values
}
\description{
Helper: mean standardize a vector
}
