% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcor.R
\name{.wcor_helper}
\alias{.wcor_helper}
\title{Helper: perform rapid unweighted and weighted correlation}
\usage{
.wcor_helper(x, w = NULL)
}
\arguments{
\item{x}{matrix of variables over which to calculate pairwise correlations}

\item{w}{vector of weights}
}
\value{
data.table with pairwise correlations
}
\description{
Helper: perform rapid unweighted and weighted correlation
}
