% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcor.R
\name{.weighted_partial_correlation}
\alias{.weighted_partial_correlation}
\title{Helper: perform rapid (parallel) weighted partial correlation calculations on a matrix}
\usage{
.weighted_partial_correlation(x, covs, covs_alt = NULL, w = NULL, ncores = 1)
}
\arguments{
\item{x}{matrix of variables over which to calculate pairwise correlations}

\item{covs}{matrix of variables that serve as covariates}

\item{covs_alt}{alternate matrix of variables that serve as covariates when pairwise correlation is not complete}

\item{w}{vector of weights}

\item{ncores}{number of cores over which to parallelize}
}
\value{
data.table with weighted pairwise partial correlations
}
\description{
Helper: perform rapid (parallel) weighted partial correlation calculations on a matrix
}
