% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extractUniqueValues}
\alias{extractUniqueValues}
\title{Extract Unique Values for Covariates from a Data Frame}
\usage{
extractUniqueValues(terms, covars)
}
\arguments{
\item{terms}{A character vector of covariate names.}

\item{covars}{A data frame containing covariates with names matching the terms.}
}
\value{
A named list where each element contains the unique values of the corresponding covariate.
}
\description{
This function takes a vector of covariate names and a data frame, and returns
the unique values for each covariate. This is useful for determining the dimensions
of a state array for a hidden Markov model.
}
\examples{
# Example data
covars <- data.frame(LH = rep(c("Sub", "Yr", "Dead"),2),
                     loc = rep(c("Wenatchee", "McNary", "John Day"),2))

# Extract unique values for LH and loc
unique_values <- extractUniqueValues(c("LH", "loc"), covars)

#Print the unique values
print(unique_values)
}
