% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcor.R
\name{\%dopar\%}
\alias{\%dopar\%}
\title{Export foreach infix operator for use within package}
\usage{
obj \%dopar\% ex
}
\arguments{
\item{obj}{left side of operator}

\item{ex}{right side of operator}
}
\description{
Export foreach infix operator for use within package
}
