% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm.R
\docType{data}
\name{icd10cm}
\alias{icd10cm}
\title{ICD10CM code description}
\format{
## `icd10cm`
A data frame with 98,583 rows and 3 columns:
\describe{
  \item{code}{ICD-10-CM code}
  \item{description}{Description of code}
  \item{vocabulary_id}{Incidates that the code is from the ICD-10-CM vocabulary}
  ...
}
}
\source{
ICD-10-CM from Athena ODHSI downloaded from https://athena.ohdsi.org/ on August 9, 2023.
}
\usage{
icd10cm
}
\description{
A table listing ICD-10-CM codes and their corresponding descriptions.
Last manually curated on August 18, 2023.
}
\keyword{datasets}
