% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_desc.R
\name{icd_desc}
\alias{icd_desc}
\title{Quickly attach ICD-10 and ICD-9/ICD10-CM codes descriptions to a data.table}
\usage{
icd_desc(
  x,
  code_col = "code",
  vocab_col = "vocabulary_id",
  simplify = FALSE,
  keep_old = NULL
)
}
\arguments{
\item{x}{a data.table with phecodes to merge in descriptions and other metadata}

\item{code_col}{name of column in `x` that contains ICD codes}

\item{vocab_col}{name of column in `x` that contains ICD vocabulary (ICD9CM, ICD10, or ICD10CM)}

\item{simplify}{if TRUE, only return `code_col`, `vocabulary_id`, and `description` columns}

\item{keep_old}{if not NULL, keep additional columns from `x`}
}
\value{
data.table with ICD code descriptions
}
\description{
Quickly attach ICD-10 and ICD-9/ICD10-CM codes descriptions to a data.table
}
\examples{
\dontrun{
icd <- data.frame(
  vocabulary_id = c("ICD9CM", "ICD10CM"),
  code = c("185", "F32.A")
)
icd_desc(icd)
}
}
