% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{ipw}
\alias{ipw}
\title{estimate inverse probability weights based on Beesley & Mukhherjee}
\usage{
ipw(
  stacked_data,
  weight_outcome_var = "WTFA_A",
  samp_var = "samp_WTFA_A",
  external_dataset = "NHIS",
  dataset_name = "MGI",
  id_var = "id",
  cancer_factor = FALSE,
  cancer_factor_var = "cancer",
  covs = c("age_50", "female", "nhw", "hypertension", "diabetes", "cancer", "anxiety",
    "depression", "bmi_cat"),
  chop = TRUE
)
}
\arguments{
\item{stacked_data}{data.table of stacked data}

\item{weight_outcome_var}{variable name of sampling weights in external dataset}

\item{samp_var}{variable name of sampling weights in internal dataset}

\item{external_dataset}{name of external dataset}

\item{dataset_name}{name of internal dataset}

\item{id_var}{variable name of id}

\item{cancer_factor}{logical, whether to include cancer factor}

\item{cancer_factor_var}{variable name of cancer factor}

\item{covs}{vector of covariates to include in model}

\item{chop}{logical, whether to chop weights}
}
\value{
return table of id and inverse probability weights
}
\description{
estimate inverse probability weights based on Beesley & Mukhherjee
}
