% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_analysis.R
\name{logistf_analysis}
\alias{logistf_analysis}
\title{Helper: function to run a binary analysis using logistf}
\usage{
logistf_analysis(
  data,
  outcome,
  exposure,
  covariates = NULL,
  logistf_pl = FALSE,
  weight_var = NULL
)
}
\arguments{
\item{data}{data frame}

\item{outcome}{name of outcome variable}

\item{exposure}{name of exposure variable}

\item{covariates}{vector of covariate names}

\item{logistf_pl}{whether to use profile likelihood in logistf}

\item{weight_var}{name of weight variable}
}
\value{
list of results from model for the exposure
}
\description{
Helper: function to run a binary analysis using logistf
}
