% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_analysis.R
\name{map_phewas}
\alias{map_phewas}
\title{function to map a binary analysis by vector of exposures using a method specified by the method argument}
\usage{
map_phewas(
  data = NULL,
  design = NULL,
  outcome,
  exposures,
  covariates,
  method,
  logistf_pl = FALSE,
  .weight_var = NULL,
  workers = 1,
  plan_strategy = future::multicore
)
}
\arguments{
\item{data}{data frame}

\item{design}{survey design object}

\item{outcome}{name of outcome variable}

\item{exposures}{vector of exposure variables in data}

\item{covariates}{vector of covariate names}

\item{method}{method to use for analysis}

\item{logistf_pl}{whether to use profile likelihood in logistf}

\item{.weight_var}{name of weight variable}

\item{workers}{number of workers to use for parallelization}

\item{plan_strategy}{strategy for parallelization}
}
\value{
return table of results from model for the exposures
}
\description{
function to map a binary analysis by vector of exposures using a method specified by the method argument
}
