% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_list.R
\name{merge_list}
\alias{merge_list}
\title{Quickly merge across a list using dplyr-based merge functions}
\usage{
merge_list(data_list, join_fn = NULL, by_var = "id", verbose = FALSE)
}
\arguments{
\item{data_list}{a list object of data frames to merge}

\item{join_fn}{the function on which to join (default = dplyr::left_join)}

\item{by_var}{the name of the variable(s) on which to join}

\item{verbose}{logical; report merging variable and merge function}
}
\value{
a merged table across objects in list ...
}
\description{
Quickly merge across a list using dplyr-based merge functions
}
\examples{
\dontrun{
data <- data.frame(
  id = letters[1:10],
  a  = 1:10,
  b  = 11:20,
  c  = 21:30
)
data_list <- list()
for (i in 2:ncol(data)) {
  data_list[[i - 1]] <- data[, c(1, i)]
}
merged <- merge_list(data_list, verbose = TRUE)
merged
}
}
