% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{MStmb,tmb_list-method}
\alias{MStmb,tmb_list-method}
\title{Estimate}
\usage{
\S4method{MStmb}{tmb_list}(object)
}
\arguments{
\item{object}{A tmb_list object.}
}
\value{
The estimate tmb object
}
\description{
Build the tmb object and estimate that parameters of the model
}
\examples{
# Define MR_settings (as empty list here; customize as needed)
MR_settings_example <- list()

# Create an instance of the tmb_list class using the WEN data set
input <- new("tmb_list",
      MR_settings = MR_settings_example,
      data = WEN)

# Run the model
fit <- MStmb(input)

}
