% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{parseFormula}
\alias{parseFormula}
\title{Parse a Formula to Extract Covariate Terms}
\usage{
parseFormula(formula)
}
\arguments{
\item{formula}{A formula object representing the covariates.}
}
\value{
A character vector of the terms (covariates) in the formula.
}
\description{
This function extracts the terms (variables) from a formula object. It is used to
identify which covariates are involved in creating a state array for a hidden Markov model.
}
\examples{
# Example formula
frm <- formula('~ LH + loc')

# Extract terms
ftrms <- parseFormula(frm)

#print terms of formula
print(ftrms)

}
