% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbfor.R
\name{pbfor}
\alias{pbfor}
\title{For loop function with lapply syntax and pretty progress bar}
\usage{
pbfor(X, FUN, pbname = NULL, simplify = FALSE, ...)
}
\arguments{
\item{X}{Object to iterate over}

\item{FUN}{Function to apply to object X}

\item{pbname}{Name of progress bar (optional)}

\item{simplify}{Try simplifying output (TRUE/FALSE)}

\item{...}{Additional arguments to pass to FUN}
}
\value{
Returns output from iteration
}
\description{
For loop function with lapply syntax and pretty progress bar
}
