% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe_desc.R
\name{phe_desc}
\alias{phe_desc}
\title{Quickly attanch phecode descriptions to a data.table}
\usage{
phe_desc(
  x,
  phe_col = "phecode",
  simplify = FALSE,
  group = TRUE,
  keep_old = NULL
)
}
\arguments{
\item{x}{a data.table with phecodes to merge in descriptions and other metadata}

\item{phe_col}{name of column in `x` that contains phecodes}

\item{simplify}{if TRUE, only return `phe_col` and `description` columns}

\item{group}{if TRUE, return `group` column}

\item{keep_old}{if not NULL, keep additional columns from `x`}
}
\value{
data.table with phecode descriptions
}
\description{
Quickly attanch phecode descriptions to a data.table
}
\examples{
\dontrun{
phecode_sub <- ms::pheinfo[1:10, .(phecode)]
phe_desc(phecode_sub)
}
}
