% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phecodex_rollup.R
\docType{data}
\name{phecodex_rollup}
\alias{phecodex_rollup}
\title{PhenomeX phenome structure}
\format{
## `phecodex_rollup`
A data frame with 7,956 rows and 2 columns:
\describe{
  \item{code}{Primary phecode label}
  \item{phecode_unrolled}{A phecode that is implied by the primary phecode label}
  ...
}
}
\source{
Unedited, raw version of phecodeX_R_rollup_map.csv available via the phecodeX github page (https://github.com/PheWAS/PhecodeX). 2023-09-21.
}
\usage{
phecodex_rollup
}
\description{
An unedited table of PheWAS X Codes (or phecodes) and their "roll up" structure.
Downloaded from github.com/PheWAS/phecodex/ on 2023-09-21.
}
\keyword{datasets}
