% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_analysis.R
\name{phewas_analysis}
\alias{phewas_analysis}
\title{function to run a binary analysis using a method specified by the method argument}
\usage{
phewas_analysis(
  data = NULL,
  design = NULL,
  outcome,
  exposure,
  covariates,
  method,
  logistf_pl = FALSE,
  .weight_var = NULL,
  verbose = TRUE,
  show_error = FALSE
)
}
\arguments{
\item{data}{data frame}

\item{design}{survey design object}

\item{outcome}{name of outcome variable}

\item{exposure}{name of exposure variable}

\item{covariates}{vector of covariate names}

\item{method}{method to use for analysis}

\item{logistf_pl}{whether to use profile likelihood in logistf}

\item{.weight_var}{name of weight variable}

\item{verbose}{whether to print offending exposure}

\item{show_error}{whether to show error messages}
}
\value{
list of results from model for the exposure
}
\description{
function to run a binary analysis using a method specified by the method argument
}
