% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phewasx.R
\name{plot_phewasx}
\alias{plot_phewasx}
\title{Plot PhecodeX PheWAS results}
\usage{
plot_phewasx(
  data,
  beta_var = "beta",
  log10p_var = "log10p",
  color_var = "color",
  phe_var = "exposure",
  threshold_color = "red",
  order_reset = TRUE,
  group_space = 20,
  label_top = 5,
  title = NULL,
  genetic_offset = 15,
  color_dot_pt = 15,
  color_dot_symbol = "●"
)
}
\arguments{
\item{data}{data of PheWAS results}

\item{beta_var}{name of column containing beta values}

\item{log10p_var}{name of column containing -log10(p) values}

\item{color_var}{name of column containing color values}

\item{phe_var}{name of column containing phecode values}

\item{threshold_color}{color of dashed line at p = 0.05}

\item{order_reset}{reset ordering of phecodes}

\item{group_space}{space between groups}

\item{label_top}{number of top phecodes to label}

\item{title}{title of plot}

\item{genetic_offset}{offset for genetic phecodes}

\item{color_dot_pt}{size of color dots}

\item{color_dot_symbol}{symbol for color dots}
}
\value{
a Manhattan plot of PheWAS results
}
\description{
Plot PhecodeX PheWAS results
}
\examples{
\dontrun{
## TO DO
}
}
