% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psw.R
\name{psw}
\alias{psw}
\title{estimate poststratification weights based on Beesley & Mukhherjee}
\usage{
psw(
  int_data,
  ext_data = NULL,
  id_var = "id",
  age_var = "AgeLastEntry",
  age_bin = FALSE,
  age_bin_breaks = c(0, 18, 35, 65, 80, 150),
  covs = c("age_bin", "cad", "smoker", "diabetes", "female"),
  not_num_vars = NULL,
  psu_var = "psu_nhanes",
  strata_var = "strata_nhanes",
  weight_var = "weight_nhanes",
  chop = TRUE
)
}
\arguments{
\item{int_data}{internal data to use for estimating weights}

\item{ext_data}{external data to use for estimating weights}

\item{id_var}{variable name for id}

\item{age_var}{variable name for age}

\item{age_bin}{logical, whether to bin age}

\item{age_bin_breaks}{breaks for age binning}

\item{covs}{covariates to use for estimating weights}

\item{not_num_vars}{variables to exclude from numeric covariates}

\item{psu_var}{variable name for psu}

\item{strata_var}{variable name for strata}

\item{weight_var}{variable name for weight}

\item{chop}{logical, whether to truncate weights}
}
\value{
return data.table with id and poststratification weights
}
\description{
estimate poststratification weights based on Beesley & Mukhherjee
}
