% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_ms}
\alias{scale_color_ms}
\title{Maxwell Salvatore's custom ggplot2 color palette}
\usage{
scale_color_ms(black_first = FALSE, ...)
}
\arguments{
\item{black_first}{Should black be the first color in the palette (TRUE/FALSE)}

\item{...}{Additional arguments}
}
\value{
A discrete ggplot2 color scale
}
\description{
Maxwell Salvatore's custom ggplot2 color palette
}
