% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_variables.R
\name{standardize_variables}
\alias{standardize_variables}
\title{Rapidly standarize variables in a datasset}
\usage{
standardize_variables(data, cols = NULL, all_numeric = FALSE)
}
\arguments{
\item{data}{data.table or data.frame of data}

\item{cols}{vector of column names or indices to standardize}

\item{all_numeric}{logical; standardize all numeric variables; will override `cols`}
}
\value{
data.table of data with standardized variables
}
\description{
Rapidly standarize variables in a datasset
}
\examples{
\dontrun{
data <- data.frame(
a = rnorm(10000, mean = 10, sd = 2),
b = rnorm(10000, mean = 10, sd = 2),
c = as.character(rnorm(10000, mean = 10, sd = 2)),
d = rnorm(10000, mean = 10, sd = 2)
)

standardize_variables(data = data, all_numeric = TRUE)
}
}
