% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_analysis.R
\name{svyglm_analysis}
\alias{svyglm_analysis}
\title{Helper: function to run a binary analysis using survey::svyglm}
\usage{
svyglm_analysis(design, outcome, exposure, covariates = NULL)
}
\arguments{
\item{design}{survey design object}

\item{outcome}{name of outcome variable}

\item{exposure}{name of exposure variable}

\item{covariates}{vector of covariate names}
}
\value{
list of results from model for the exposure
}
\description{
Helper: function to run a binary analysis using survey::svyglm
}
