% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ms}
\alias{theme_ms}
\title{Maxwell Salvatore's custom ggplot2 theme}
\usage{
theme_ms(
  show_grid_lines = TRUE,
  show_axis_titles = TRUE,
  show_legend_title = TRUE,
  font_family = NULL
)
}
\arguments{
\item{show_grid_lines}{Should grid lines be shown (TRUE/FALSE)}

\item{show_axis_titles}{Should axis titles be shown (TRUE/FALSE)}

\item{show_legend_title}{Should axis title by shown (TRUE/FALSE)}

\item{font_family}{Font to use in plot (Google Font, default = 'Public Sans')}
}
\value{
A complete ggplot2 theme
}
\description{
Maxwell Salvatore's custom ggplot2 theme
}
