% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_forest.R
\name{tidy_forest}
\alias{tidy_forest}
\title{Conduct a tidy random forest analysis with hyperparameter tuning via tidymodels framework}
\usage{
tidy_forest(
  dataset,
  outcome_var,
  drop_vars = NULL,
  split_prop = 0.5,
  num_threads = 6,
  importance = "permutation",
  mode = "classification",
  levels = 5,
  best_metric = "accuracy"
)
}
\arguments{
\item{dataset}{A data frame}

\item{outcome_var}{A string representing outcome variable}

\item{drop_vars}{A vector of strings representing variables to drop}

\item{split_prop}{A number representing proportion of data to use for training}

\item{num_threads}{A number representing number of threads to use}

\item{importance}{A string representing importance method}

\item{mode}{A string representing mode}

\item{levels}{A number representing number of levels}

\item{best_metric}{A string representing best metric}
}
\value{
A list of objects
}
\description{
Conduct a tidy random forest analysis with hyperparameter tuning via tidymodels framework
}
