% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{tmb_list-class}
\alias{tmb_list-class}
\title{tmb_list Class}
\description{
This class holds the data needed for TMB-based mark-recapture analysis.
}
\section{Slots}{

\describe{
\item{\code{MR_settings}}{A list of mark-recapture settings, expected to contain ...}

\item{\code{data}}{A list of TMB data. Must contain \code{id}, \code{loc}, and \code{n} columns in its data frames.}
}}

