% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_analysis.R
\name{weighted_analysis}
\alias{weighted_analysis}
\title{Helper: function to run a weighted binary analysis using svyglm unless there's a separation concern, then run weighted logsitf}
\usage{
weighted_analysis(
  data,
  design,
  outcome,
  exposure,
  covariates,
  method = "svyglm",
  logistf_pl = FALSE,
  weight_var = NULL
)
}
\arguments{
\item{data}{data frame}

\item{design}{survey design object}

\item{outcome}{name of outcome variable}

\item{exposure}{name of exposure variable}

\item{covariates}{vector of covariate names}

\item{method}{method to use for analysis}

\item{logistf_pl}{whether to use profile likelihood in logistf}

\item{weight_var}{name of weight variable}
}
\value{
list of results from model for the exposure
}
\description{
Helper: function to run a weighted binary analysis using svyglm unless there's a separation concern, then run weighted logsitf
}
