% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Print data.table as tibbles for viewing}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Currently unused}

\item{pkgname}{Currently unused}
}
\value{
Rewrites print.data.table in background
}
\description{
Print data.table as tibbles for viewing
}
