% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence_analysis.R
\name{.quick_cooccur_mod}
\alias{.quick_cooccur_mod}
\title{Helper: perform rapid unweighted and weighted cooccurrence analysis}
\usage{
.quick_cooccur_mod(
  data,
  covariates = c("age_at_threshold", "female", "length_followup"),
  outcome = "case",
  exposure = "X157",
  weight_var = NULL,
  weight_dsn = NULL,
  evalue = TRUE,
  logistf_pl = TRUE,
  check_separation = TRUE,
  detect_separation = TRUE
)
}
\arguments{
\item{data}{dataset}

\item{covariates}{vector of covariates}

\item{outcome}{outcome variable}

\item{exposure}{exposure variable}

\item{weight_var}{weight variable name}

\item{weight_dsn}{survey design object}

\item{evalue}{logical; calculate evalues (default = TRUE)}

\item{logistf_pl}{logical; use profile likelihood (TRUE) for confidence intervals and tests or Wald (FALSE) in logistf; (default = TRUE)}

\item{check_separation}{logical; check for separation (default = TRUE)}

\item{detect_separation}{logical; use detect_separation (TRUE) or skip and use logistf regardless (FALSE); (default = TRUE)}
}
\value{
data.table with results from cooccurrence analysis
}
\description{
Helper: perform rapid unweighted and weighted cooccurrence analysis
}
