% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcor.R
\name{fcor}
\alias{fcor}
\title{Perform rapid (parallel) unweighted and weighted (partial) correlations on a matrix}
\usage{
fcor(x, covs = NULL, covs_alt = NULL, w = NULL, n_cores = 1, verbose = FALSE)
}
\arguments{
\item{x}{matrix of variables over which to calculate pairwise correlations}

\item{covs}{matrix of variables that serve as covariates}

\item{covs_alt}{alternate matrix of variables that serve as covariates when pairwise correlation is not complete}

\item{w}{vector of weights}

\item{n_cores}{number of cores over which to parallelize}

\item{verbose}{logical; report what correlations are being calculate}
}
\value{
data.table of pairwise correlations
}
\description{
Perform rapid (parallel) unweighted and weighted (partial) correlations on a matrix
}
\examples{
\dontrun{
data        <- data.frame(matrix(rnorm(100 * 50), ncol = 50))
covs        <- data.frame(matrix(rnorm(100 * 2), ncol = 2))
names(covs) <- c("C1", "C2")
wgt <- 1/runif(100)

fcor(x = data, covs = covs, w = wgt, n_cores = 1, verbose = TRUE)
}
}
